Refactor requests and headers into separate modules
