+++ request.py
@@ -1,0 +1,96 @@
+import io
+import socket
+import typing
+
+from headers import Headers
+
+
+class BodyReader(io.IOBase):
+    def __init__(self, sock: socket.socket, *, buff: bytes = b"", bufsize: int = 16_384) -> None:
+        self._sock = sock
+        self._buff = buff
+        self._bufsize = bufsize
+
+    def readable(self) -> bool:
+        return True
+
+    def read(self, n: int) -> bytes:
+        """Read up to n number of bytes from the request body.
+        """
+        while len(self._buff) < n:
+            data = self._sock.recv(self._bufsize)
+            if not data:
+                break
+
+            self._buff += data
+
+        res, self._buff = self._buff[:n], self._buff[n:]
+        return res
+
+
+class Request(typing.NamedTuple):
+    method: str
+    path: str
+    headers: Headers
+    body: BodyReader
+
+    @classmethod
+    def from_socket(cls, sock: socket.socket) -> "Request":
+        """Read and parse the request from a socket object.
+
+        Raises:
+          ValueError: When the request cannot be parsed.
+        """
+        lines = iter_lines(sock)
+
+        try:
+            request_line = next(lines).decode("ascii")
+        except StopIteration:
+            raise ValueError("Request line missing.")
+
+        try:
+            method, path, _ = request_line.split(" ")
+        except ValueError:
+            raise ValueError(f"Malformed request line {request_line!r}.")
+
+        headers = Headers()
+        buff = b""
+        while True:
+            try:
+                line = next(lines)
+            except StopIteration as e:
+                # StopIteration.value contains the return value of the generator.
+                buff = e.value
+                break
+
+            try:
+                name, _, value = line.decode("ascii").partition(":")
+                headers.add(name, value.lstrip())
+            except ValueError:
+                raise ValueError(f"Malformed header line {line!r}.")
+        body = BodyReader(sock, buff=buff)
+        return cls(method=method.upper(), path=path, headers=headers, body=body)
+
+
+def iter_lines(sock: socket.socket, bufsize: int = 16_384) -> typing.Generator[bytes, None, bytes]:
+    """Given a socket, read all the individual CRLF-separated lines
+    and yield each one until an empty one is found.  Returns the
+    remainder after the empty line.
+    """
+    buff = b""
+    while True:
+        data = sock.recv(bufsize)
+        if not data:
+            return b""
+
+        buff += data
+        while True:
+            try:
+                i = buff.index(b"\r\n")
+                line, buff = buff[:i], buff[i + 2:]
+                if not line:
+                    return buff
+
+                yield line
+            except IndexError:
+                break
